+++ cmd/hidden_lake/service/internal/config/init.go
@@ -31,6 +31,7 @@ func InitConfig(cfgPath string, initCfg *SConfig) (IConfig, error) {
 				FTCP:  hls_settings.CDefaultTCPAddress,
 				FHTTP: hls_settings.CDefaultHTTPAddress,
 			},
+			FF2FDisabled: hls_settings.CDefaultF2FDisabled,
 			FServices: map[string]*SService{
 				hlm_settings.CTitlePattern: {FHost: hls_settings.CDefaultServiceHLMAddress},
 			},