+++ cmd/hidden_lake/service/internal/handler/network_online_test.go
@@ -126,7 +126,7 @@ func testOnlinePushNode(cfgPath, dbPath string) (anonymity.INode, context.Cancel
 	node.HandleFunc(
 		pkg_settings.CServiceMask,
 		HandleServiceTCP(
-			cfg,
+			config.NewWrapper(cfg),
 			logger.NewLogger(
 				logger.NewSettings(&logger.SSettings{}),
 				func(_ logger.ILogArg) string { return "" },