+++ cmd/hidden_lake/service/pkg/app/init_pprof.go
@@ -3,5 +3,5 @@ package app
 import "github.com/number571/go-peer/internal/pprof"
 
 func (p *sApp) initServicePPROF() {
-	p.fServicePPROF = pprof.InitPprofService(p.fConfig.GetAddress().GetPPROF())
+	p.fServicePPROF = pprof.InitPprofService(p.fCfgW.GetConfig().GetAddress().GetPPROF())
 }