+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 *??? ??, ????*
 
+### IMPROVEMENTS
+
+- Update `pkg/network/anonymity|cmd/hidden_lake/service`: append f2f_disable option
+
 ### CHANGES
 
 - Update `pkg/crypto/asymmetric`: IAddress -> hashing.IHasher